Pharmacy warehouse
Pharmacy warehouse warehouse will contain a table with 50 positions for the products and a pointer that will point to the next available position in the table to insert a new product. The array numbering will start at 0 and increment by 1 each time a product is added. When a product is deleted, the table position will remain empty and the message "DELETED" will be displayed. When a product is updated, the user will give the original product to be replaced with the new product to be introduced. If a product is searched for, its position in the table will be returned, otherwise -1 if not found